+++ config/page.js
@@ -1,0 +1,5 @@
+module.exports = {
+  title: 'lovli.js | Hello World',
+  port: 3000,
+  token_secret: 'thisislovlijs'
+};