+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+### v0.3.0
+- added production configuration
+  - run ```npm run prod``` to start the app in production mode
+  - run ```npm run build``` to generate the combined js files
+  - run ```node .build/server.bundle.js``` to start your app manually
+- added page configuration in ```config/page.js```
+  - there you can change title, port and horizon token_secret
+- fixed webpack config to only include "source" folder for babel-loader
+
 ### v0.2.2
 - explicitly setting token_secret for horizon server to make sure tokens are valid on server restart
 - update @horizon/client to 0.0.4-4